Add new function, remove extraneous code
I removed the commented-out code for a chart that I am not going to
implement.  Added code to calculate the number of tools that are
missing between one and ten values, and updated the chart generating
function to take the data directly from the data object rather than
having it hard-coded.